Reference manually created hosted zone domain
The route53 hosted zone was created outside of our terraform so we are
unable to reference it as a terraform resource.

The subdomain naming convention conlflicts with our environment names so
we cannot infer the domain based on the environment.

In route53 we have:
  dev.vpn.justice.gov.uk
  prep.vpn.justice.gov.uk
  vpn.justice.gov.uk

These were added to parameter store where they can be pulled down during
build.