Revert changing domain name to CNAME
This changeset is invalid even when trying to remove the A record in the
console. This might be a bug in AWS, needs more research, current error
message is:

"Bad request.
(InvalidInput 400: Invalid request: Expected exactly one of [AliasTarget, all of [TTL, and ResourceRecords], or TrafficPolicyInstanceId], but found none in Change with [Action=DELETE, Name=dhcp-lease-db-development.dev.vpn.justice.gov.uk., Type=A, SetIdentifier=eu-west-2])"