Output Terraform outputs object
This will be the payload of configurations created by Terraform that
other applications depend on.

An example of this is the docker container image building which runs in
a separate pipeline. The building and deploying of the image requires
values created in Terraform which may change over time.

This entire output object will be stored in SSM parameter store and can
be pulled out to get the most up to date values created by Terraform.