Manually remove auto scaling groups before destroy
Due to a bug in Terraform, ECS is unable to delete before the auto
scaling group has been removed.

Use the aws command line in combination with your current workspace to
delete the auto scaling group as a separate step before running
terraform destroy.

This is wrapped up in `make destroy`, and `terraform destroy` should not
be used.

Because calling aws from the command line is unable to assume a role
unless the arn is known, the `aws-vault` commands need to be hardcoded
within the Makefile.