Upgrade Terraform to version 0.13 (#26)
We are seeing dependency issues when running `terraform destroy`
Two issues are preventing a clean destroy:

1. Terraform attempts to destroy network resources before other
resources. This fails because you cannot destroy a VPC when you have
services running in it.

2. Terraform attempts to destroy the ECS cluster before the auto scaling
group that serves as the compute for the capacity provider.

This PR addresses the first issue, by leveraging the module `depends_on`
feature in Terraform 0.13.

The second issue still needs to be addressed by extracting the auto
scaling group into its own module and having the ECS cluster depend on
it. https://github.com/terraform-providers/terraform-provider-aws/issues/4852

To use this for local development, run `make init`, which will
reconfigure the state to use the new version of Terraform.

A PR following this will remove the `-reconfigure` flag from the
Makefile once everyone has upgraded.