Manually remove auto scaling groups before destroy (#27)
* Upgrade Terraform to version 0.13

We are seeing dependency issues when running `terraform destroy`
Two issues are preventing a clean destroy:

1. Terraform attempts to destroy network resources before other
resources. This fails because you cannot destroy a VPC when you have
services running in it.

2. Terraform attempts to destroy the ECS cluster before the auto scaling
group that serves as the compute for the capacity provider.

This PR addresses the first issue, by leveraging the module `depends_on`
feature in Terraform 0.13.

The second issue still needs to be addressed by extracting the auto
scaling group into its own module and having the ECS cluster depend on
it. https://github.com/terraform-providers/terraform-provider-aws/issues/4852

To use this for local development, run `make init`, which will
reconfigure the state to use the new version of Terraform.

A PR following this will remove the `-reconfigure` flag from the
Makefile once everyone has upgraded.

* Manually remove auto scaling groups before destroy

Due to a bug in Terraform, ECS is unable to delete before the auto
scaling group has been removed.

Use the aws command line in combination with your current workspace to
delete the auto scaling group as a separate step before running
terraform destroy.

This is wrapped up in `make destroy`, and `terraform destroy` should not
be used.

Because calling aws from the command line is unable to assume a role
unless the arn is known, the `aws-vault` commands need to be hardcoded
within the Makefile.