Allow listening on the traffic port for health checks
We use dynamic port mapping with ECS to allow more than one container to
run on an EC2 instance. The LB health check is only able to health check
the traffic port which currently only listens on UDP.

Map port 67 for TCP to the container which will have Nginx listening on
this port.