Remove active health check on target group
We have implemented sophisticated KEA perfdhcp health checks and don't
use this health check as validation that the node is healthy.

This NLB health check takes around a minute to finish and even though
ECS reports that the server is up, it still needs to wait for this
health checks to complete before being added to the load balancer.

By removing this health check, we still get a required passive health check
before target is consisedered healthy.