Make production bastion a temporary piece of infrastructure
This bastion server should only exist when access is needed to access the VMs
in Corsham. This improves the security posture of the DHCP service.

Tests in Corsham are now automated and run on an infinite loop, so no
need to run manually.

Only for debugging the remote VM would we spin up a bastion.

Currently this is setting to create this instance is set to true for testing.

A future commit will remove this.