Refactor Admin IAM policies
There are two roles that need to be assumed by ECS:
1. Task Execution role
2. Task role (application)

The execution role relies on a predefined AWS IAM role to allow it to
pull from ECR and write to CloudWatch.

The task role needs to access the ECS Service for DHCP to do a rolling
deployment. It also needs access to S3 to write the config files.