Automatically set retention period of all CloudWatch logs
Run a script at the end of a build to ensure all CloudWatch logs have a
retention period of 7 days. This is to ensure we are inline with DPIA
requirements. These logs will be instantly shipped to the security
logging endpoint, and keeping them for 7 days is a safety precaution in
case something goes wrong with the shippers.