Cloudwatch log retention (#54)
* Refactor DNS / DHCP IAM policies to restrict access

This refines the task and execution roles to adhere to the principle of
least priviledge.

This introduces some duplication, where features from IAM and S3 have
been moved from dns_dhcp_common into the dns and dhcp modules.

This is unavoidable and Terraform will end up with a cycle error if
calling a child module.

A future task will look to remove this duplication by introducing an IAM
module which can be shared by admin, dns and dhcp.

* Automatically set retention period of all CloudWatch logs

Run a script at the end of a build to ensure all CloudWatch logs have a
retention period of 7 days. This is to ensure we are inline with DPIA
requirements. These logs will be instantly shipped to the security
logging endpoint, and keeping them for 7 days is a safety precaution in
case something goes wrong with the shippers.