Add Transit Gateway as the default route for DHCP and DNS
Any unrouted requests need to go through to the transit gateway to
handle traffic from the network. DNS will get a NAT gateway in a future
pull request to add a known source IP to it that can be whitelisted.

Recalculate CIDR subnets and remove one of the 3 AZs, running in 2 is good enough.
Set task count for DHCP to 1 to prevent any concurrency issues when
traffic is high.

This requires a rebuild of all the infrastructure as you cannot modify
the network in place.