Boot the API as standby not primary
The API (used to validate configurations and list leases to the admin)
is currently booting as a primary server in the peer relationship.
This means that the API is sending heartbeats to the standby server and
standby won't ever consider primary as down.

Removing the HA config from the API would require a copy of all the
underlying ECS infrastructure and a separate configuration file.

By booting the API as standby, it won't interfere with the 2 peers when
detecting if they are up or not.