Set sql_mode to STRICT_ALL_TABLES
Without strict mode the server tries to do proceed with the action when an error
might have been a more secure choice. For example, by default MySQL will
truncate data if it does not fit in a field, which can lead to unknown
behaviour, or be leveraged by an attacker to circumvent data validation.