Create DHCP heartbeat instance
This will be used to as a DHCP client to constantly lease IPs from the
DHCP service.

The result of this lease request will be logged to a Cloudwatch
log group.

A script (to be implemented) from CodePipeline will poll this log group to make sure the
heartbeat is still able to lease IPs.

If the heartbeat fails to lease an IP, the pipeline will abort and won't
deploy to the next environment (pre-prod, production).