Create routes to test VMs from public subnets
To gain access to the test VMs in Corsham and Model office, we have
to go through a jumpbox residing in our VPC.

The bastion server needs to be in a public subnet to allow SSH access to
it. The public subnets have an IGW as the default route.

Create explicit routes for Corsham and Model office to point to the
destination IPs.