Reverted to commit id: 2a6103a9ead211cb25611d8d973ce6854b41750f
Fixes unecessary changes to cognito_identity
The "appid" is appended to the URL as a query string, this doesn't
appear to be stored with the amended SLORedirectBindingURI or
SSORedirectBindingURI hence Terraform attempts to change it on every
plan or apply.
By removing the query string from the value Terraform stops reporting
changes.