Refactored label module to accept owner_email
Previously the owner email was not defined in the tags, as per MOJ
mandatory tags documentation:
https://technical-guidance.service.justice.gov.uk/documentation/standards/documenting-infrastructure-owners.html#mandatory
We need to have the owner email set.

During the refectoring process, we found that the owner_email was also
set incorrectly in the buildspec files, we have removed this value from
these file and set them within the terraform instead.