Removing deprecated TF_VARS script
The TF_VARS stored in a single AWS SSM Parameter was drifting from the
actual values stored in the individual AWS SSM Parameters.
A new script had been written to pull the same paramters as specified in
the buildspec.yml as used by the CI (AWS CodePipeline).
This is the consistent way to generate the TV_VARS required.