Added in new jumpbox module
We needed an EC2 instance to perform load testing. Upon investigation,
the corsham_test_bastion module already in place did not have AWS SSM (session manager) enabled.
The benefit to SSM is it makes the EC2 instance more secure, as it does
not require a public IP and exposed SSH ports.
We did bring in a bastion module from staff-infrastructure-network-services,
We then tried to depricate the corsham module in favour of the new
module, however we found the corsham module has unquie use case. Rather
than investigating the corsham module we opted to leave it as it is, and
add our new module to enable our load testing. A ticket will be created
to investigate, the corsham module and make a updates / refactoring if
required.