Added conditional SSM vpc endpoints
We currently have a switch to be able to turn on or off the load testing
ec2. We decide it was ineffienct to have VPC endpoints being deployed
and not being used when the load testing bastion is off. Thus we have
added a condition for this. We have also refectored the naming of the
module to be more reflective of the load testing use case.